+++ extensions/iceworks-component-builder/package.json
@@ -102,9 +102,11 @@
   "dependencies": {
     "@alifd/next": "^1.19.4",
     "@alifd/theme-design-pro": "^0.x",
-    "@iceworks/material-service": "^0.1.0-beta-2",
-    "@iceworks/material-ui": "^0.1.0-beta-2",
-    "@iceworks/vscode-webview": "^0.1.0-beta-2",
+    "@iceworks/block-service": "^0.1.0",
+    "@iceworks/material-service": "^0.1.0",
+    "@iceworks/project-service": "^0.1.0",
+    "@iceworks/material-ui": "^0.1.0",
+    "@iceworks/vscode-webview": "^0.1.0",
     "moment": "^2.24.0",
     "react": "^16.4.1",
     "react-dom": "^16.4.1"
