+++ extensions/iceworks-component-builder/src-extension/extension.ts
@@ -10,7 +10,7 @@ export function activate(context: vscode.ExtensionContext) {
   console.log('Congratulations, your extension "iceworks-component-builder" is now active!');
 
   function activeWebview() {
-    const webviewPanel: vscode.WebviewPanel = window.createWebviewPanel('iceworks', 'Create Page', ViewColumn.One, {
+    const webviewPanel: vscode.WebviewPanel = window.createWebviewPanel('iceworks', 'Create Component', ViewColumn.One, {
       enableScripts: true,
       retainContextWhenHidden: true,
     });
