+++ extensions/iceworks-component-builder/src-extension/services/block.ts
@@ -1,0 +1,6 @@
+import { bulkCreate as originBulckCreate } from '@iceworks/block-service';
+import { componentsPath } from '@iceworks/project-service';
+
+export const bulkGenerate = async function(blocks: any) {
+  return await originBulckCreate(blocks, componentsPath);
+}
