+++ extensions/iceworks-component-builder/src-extension/services/index.ts
@@ -1,3 +1,4 @@
 import * as material from './material';
+import * as block from './block';
 
-export default { material };
+export default { material, block };
