+++ extensions/iceworks-page-builder/build.json
@@ -23,6 +23,7 @@
       {
         "activeInDev": true
       }
-    ]
+    ],
+    "./build.plugin.js"
   ]
 }
