+++ extensions/iceworks-page-builder/build.plugin.js
@@ -1,0 +1,8 @@
+module.exports = ({onGetWebpackConfig}) => {
+  onGetWebpackConfig((config) => {
+    config.node
+      .set('fs', 'empty')
+      .set('net', 'empty')
+      .set('tls', 'empty');
+  });
+}
