+++ packages/block-service/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@iceworks/page-service",
+  "name": "@iceworks/block-service",
   "version": "0.1.0",
-  "description": "iceworks page service for VSCode extension.",
+  "description": "iceworks block service for VSCode extension.",
   "files": [
     "lib"
   ],
@@ -10,11 +10,9 @@
   "dependencies": {
     "@iceworks/material-utils": "^0.1.0",
     "@iceworks/project-service": "^0.1.0",
-    "@iceworks/block-service": "^0.1.0",
-    "ejs": "^3.1.3",
-    "fs-extra": "^9.0.0",
-    "prettier": "^2.0.5",
-    "uppercamelcase": "^3.0.0"
+    "fs-extra": "^8.1.0",
+    "uppercamelcase": "^3.0.0",
+    "ice-npm-utils": "^1.4.0"
   },
   "devDependencies": {
     "@types/vscode": "^1.45.0"
