+++ packages/page-service/src/constant.ts
@@ -1,9 +1,1 @@
-import * as vscode from 'vscode';
-import * as path from 'path';
-
-const { workspace } = vscode;
-
-export const projectPath = workspace.rootPath!;
-export const pagesPath = path.join(projectPath, 'src', 'pages');
-export const componentDirName = 'components';
 export const templateFileName = 'template.jsx.ejs';
