+++ tsconfig.json
@@ -4,10 +4,12 @@
     {"path": "packages/ice-npm-utils"},
     {"path": "packages/config"},
     {"path": "packages/generate-project"},
+    {"path": "packages/project-service"},
     {"path": "packages/iceworks-cli"},
     {"path": "packages/material-utils"},
     {"path": "packages/material-service"},
+    {"path": "packages/block-service"},
     {"path": "packages/page-service"},
-    {"path": "packages/vscode-webview"}
+    {"path": "packages/vscode-webview"},
   ]
 }
